Refactor req with user methods and type definition
